refactor: request moved to searchResults.tsx, props changed
