+++ src/components/Home/Search/SearchResults/SearchResults.tsx
@@ -1,5 +1,7 @@
 import { GoLocation, Loading, SavedLocations } from 'components';
-import React, { SetStateAction, useCallback, useContext } from 'react';
+import { ENDPOINTS } from 'consts';
+import { useDebounce, useGetRequest } from 'hooks';
+import React, { useCallback, useContext } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { Dispatch } from 'redux';
@@ -12,37 +14,41 @@ import {
   ThemeContext,
   ThemeContextConfig
 } from 'store';
-import { LocationSearch } from 'types/interfaces';
+import { LocationSearch, RequestDataConfig } from 'types/interfaces';
 import styles from './SearchResults.module.scss';
 
 interface SearchResultsProps {
-  searchResults: LocationSearch;
-  setSearchTerm: React.Dispatch<SetStateAction<string>>;
-  setDisplaySearchResults: React.Dispatch<SetStateAction<boolean>>;
-  searchLoading: boolean;
   displaySearchResults: boolean;
+  searchTerm: string;
+  onResultSelected: () => void;
 }
 
 const SearchResults: React.FunctionComponent<SearchResultsProps> = ({
-  searchResults,
-  setSearchTerm,
-  setDisplaySearchResults,
-  searchLoading,
-  displaySearchResults
+  displaySearchResults,
+  searchTerm,
+  onResultSelected
 }) => {
   const { theme }: ThemeContextConfig = useContext(ThemeContext);
-  const { locations } = searchResults;
   const { currentLocation, prevSearches } = useSelector<CombinedState, LocationState>(
     state => state.location
   );
   const dispatch: Dispatch<LocationActionConfig> = useDispatch();
   const navigate = useNavigate();
+  const debouncedSearchTerm: string = useDebounce(searchTerm);
+  const {
+    data: searchResults = { locations: [] },
+    loading: searchLoading = true,
+    error = null
+  }: RequestDataConfig<LocationSearch> = useGetRequest(
+    ENDPOINTS.locationSearch,
+    debouncedSearchTerm
+  );
+  const { locations } = searchResults;
 
   const clickHandler = useCallback(
     (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
       const target = event.target as HTMLDivElement;
-      setSearchTerm('');
-      setDisplaySearchResults(false);
+      onResultSelected();
       if (target.id !== currentLocation) {
         dispatch({ type: LocationActions.SAVE_SEARCH, payload: target.id });
         navigate(`/cities/${target.id}`);
@@ -50,8 +56,9 @@ const SearchResults: React.FunctionComponent<SearchResultsProps> = ({
         navigate('/cities/current');
       }
     },
-    [event]
+    [currentLocation, dispatch, navigate, onResultSelected]
   );
+
   if (!displaySearchResults) return <></>;
   return (
     <ul className={`${styles.container} ${theme === Theme.DARK && styles[theme]}`}>
